// ui-app/src/app/models/api-entities/business-partner-profile-request-params.ts
export interface BusinessPartnerProfileRequestParams {
  ario_marketing_requested?: boolean;
  collateral_downloaded?: boolean;
  partner_training_completed?: boolean;
  sales_training_completed?: boolean;
  facebook_sharing_requested?: boolean;
  twitter_sharing_requested?: boolean;
  linkedin_sharing_requested?: boolean;
  vanity_added_to_website?: boolean;
}