// ui-app/src/app/models/tracked-object-state.ts
export enum TrackedObjectState {
  clicked = 'clicked',
  created = 'created',
  invited = 'invited',
  invoiced = 'invoiced',
  linked = 'linked',
  paid = 'paid',
  sent = 'sent',
  securityViolation = 'security_violation',
  subscribedAutoSend = 'subscribed_auto_send',
  unsubscribedAutoSend = 'unsubscribed_auto_send',
  viewed = 'viewed',
  createdFromQuickBooks = 'created_from_quickbooks',
  updatedFromQuickBooks = 'updated_from_quickbooks'
}